Rework client creation so dependent objects are created and maintained in one place
